Move logging init code into its own module
